Index

MixtureDensityNetworks.MDN

MixtureDensityNetworks.MDN

MixtureDensityNetworks.Machine

MixtureDensityNetworks.Machine

MixtureDensityNetworks.MixtureDensityNetwork

MixtureDensityNetworks.MixtureDensityNetwork

MixtureDensityNetworks.fit!

MixtureDensityNetworks.generate_data

MixtureDensityNetworks.likelihood_loss

MixtureDensityNetworks.predict

MixtureDensityNetworks.predict_mean

MixtureDensityNetworks.predict_mode

API

MixtureDensityNetworks.MDN — Type

MDN

A model type for constructing a MDN, based on MixtureDensityNetworks.jl, and implementing the MLJ model interface.

From MLJ, the type can be imported using

MDN = @load MDN pkg=MixtureDensityNetworks

Do model = MDN() to construct an instance with default hyper-parameters. Provide keyword arguments to override hyper-parameter defaults, as in MDN(mixtures=...).

A neural network which parameterizes a Gaussian Mixture Model (GMM) distributed over the target varible y conditioned on the features X.

Training Data

In MLJ or MLJBase, bind an MDN instance model to data with mach = machine(model, X, y) where

X: any table of input features (eg, a DataFrame) whose columns belong to the Continuous scitypes`.

y: the target, which can be any AbstractVector whose element scitype is Continuous.

Hyperparameters

mixtures=5: number of Gaussian mixtures in the predicted distribution

layers=[128,]: hidden layer topology, starting from the first hidden layer

η=1e-3: learning rate used for the optimizer

epochs=1: number of epochs to train the model

batchsize=32: batch size used during training

Operations

predict(mach, Xnew): return the distributions over the target conditioned on the new features Xnew having the same scitype as X above.

predict_mode(mach, Xnew): return the largest modes of the distributions over targets conditioned on the new features Xnew having the same scitype as X above.

predict_mean(mach, Xnew): return the means of the distributions over targets conditioned on the new features Xnew having the same scitype as X above.

predict_median(mach, Xnew): return the medians of the distributions over targets conditioned on the new features Xnew having the same scitype as X above.

Fitted Parameters

The fields of fitted_params(mach) are:

fitresult: the trained mixture density model, compatible with the Flux ecosystem.

Report

learning_curve: the average training loss for each epoch.

best_epoch: the epoch (starting from 1) with the lowest training loss.

best_loss: the best (lowest) loss encountered durind training. Corresponds to the average loss of the best epoch.

Accessor Functions

training_losses(mach) returns the learning curve as a vector of average training losses for each epoch.

Examples

using MLJ
MDN = @load MDN pkg=MixtureDensityNetworks
mdn = MDN(mixtures=12, epochs=100, layers=[512, 256, 128])
X, y = make_regression(100, 1) # synthetic data
mach = machine(mdn, X, y) |> fit!
Xnew, _ = make_regression(3, 1)
ŷ = predict(mach, Xnew) # new predictions
report(mach).best_epoch # best epoch encountered during training 
report(mach).best_loss # best loss encountered during training 
training_losses(mach) # learning curve

MixtureDensityNetworks.MDN — Method

MDN(; mixtures=5, layers=[128], η=1e-3, epochs=1, batchsize=32)

Defines an MDN model with the given hyperparameters.

Parameters

mixtures: The number of gaussian mixtures to use in estimating the conditional distribution (default=5).

layers: A vector indicating the number of nodes in each of the hidden layers (default=[128,]).

η: The learning rate to use when training the model (default=1e-3).

epochs: The number of epochs to train the model (default=1).

batchsize: The batchsize to use during training (default=32).

MixtureDensityNetworks.Machine — Type

mutable struct Machine

A struct for associating a model's hyperparameters with its training state.

Parameters

hypers::MDN

fitresult::Union{Nothing, MixtureDensityNetwork}

report::NamedTuple{(:learning_curve, :best_epoch, :best_loss), Tuple{Vector{Float64}, Int64, Float64}}

MixtureDensityNetworks.Machine — Method

Machine(hypers::MDN)

Binds a collection of hyperparameters with a training state for fitting, evaluating, and predicting.

Parameters

hypers: The hyperparameters we want our model to conform to.

MixtureDensityNetworks.MixtureDensityNetwork — Type

struct MixtureDensityNetwork

A custom Flux model whose predictions paramaterize a Gaussian Mixture Model.

Parameters

hidden::Flux.Chain

μ::Flux.Dense

Σ::Flux.Dense

π::Flux.Chain

MixtureDensityNetworks.MixtureDensityNetwork — Method

MixtureDensityNetwork(
    input::Int64,
    layers::Vector{Int64},
    mixtures::Int64
) -> MixtureDensityNetwork

Construct a new MixtureDensityNetwork.

Parameters

input: The dimension of the input features.

layers: The topolgy of the hidden layers, starting from the first layer.

mixtures: The number of Gaussian mixtures to use in the conditional distribution.

MixtureDensityNetworks.fit! — Method

fit!(
    machine::Machine,
    X::Matrix{<:Real},
    Y::Matrix{<:Real}
) -> NamedTuple{(:learning_curve, :best_epoch, :best_loss), Tuple{Vector{Float64}, Int64, Float64}}

Fit the model to the data given by X and Y.

Parameters

machine: The machine to be trained.

X: A dxn matrix where d is the number of features and n is the number of samples.

Y: A 1xn matrix where n is the number of samples.

MixtureDensityNetworks.generate_data — Method

generate_data(
    n_samples::Int64
) -> Tuple{Matrix{Float64}, Matrix{Float64}}

Generate some synthetic data for testing purposes.

Parameters

n_samples: The number of samples we want to generate.

Returns

The sythetic features X and labels Y as a tuple (X, Y).

MixtureDensityNetworks.likelihood_loss — Method

likelihood_loss(
    μ::Matrix{<:Real},
    σ::Matrix{<:Real},
    pi::Matrix{<:Real},
    y::Matrix{<:Real}
) -> Float64

Conpute the negative log-likelihood loss for a set of labels y under a Gaussian Mixture Model defined by the parameters μ, σ, and pi.

Parameters

μ: A mxn matrix of means where m is the number of Gaussian mixtures and n is the number of samples.

σ: A mxn matrix of standard deviations where m is the number of Gaussian mixtures and n is the number of samples.

pi: A mxn matrix of priors where m is the number of Gaussian mixtures and n is the number of samples.

y: A 1xn matrix of labels where n is the number of samples.

MixtureDensityNetworks.predict — Method

predict(
    machine::Machine,
    X::Matrix{<:Real}
) -> Vector{Distributions.MixtureModel}

Predict the full conditional distribution P(Y|X).

Parameters

machine: The machine with which we want to generate a prediction.

X: A dxn matrix where d is the number of features and n is the number of samples.

Returns

Returns a vector of Distributions.MixtureModel objects representing the conditional distribution for each sample.

MixtureDensityNetworks.predict_mean — Method

predict_mean(
    machine::Machine,
    X::Matrix{<:Real}
) -> AbstractVector

Predict the mean of the conditional distribution P(Y|X).

Parameters

machine: The machine with which we want to generate a prediction.

X: A dxn matrix where d is the number of features and n is the number of samples.

Returns

Returns a vector of real numbers representing the mean of the conditional distribution P(Y|X) for each sample.

MixtureDensityNetworks.predict_mode — Method

predict_mode(machine::Machine, X::Matrix{<:Real}) -> Any

Predict the mean of the Gaussian with the largest prior in the conditional distribution P(Y|X).

Parameters

machine: The machine with which we want to generate a prediction.

X: A dxn matrix where d is the number of features and n is the number of samples.

Returns

Returns a vector of real numbers representing the most significant mode for each sample.